Rename job "feats", streamline job/degree flow
Originally written via Gist in October 2013: While the label is somewhat
internal (the ideal target format is still PDF), the idea is to have the
resume be legible in raw LaTeX format, meaning the variables should have
names which are commonly-used to describe their relevant areas. "Feats"
is more like what you'd find in a role-playing game, not a professional
resume.

Additionally, the flow of declaring a job or a degree is now
streamlined, so that the relevant details are all declared in the
original command opening the job/degree section. It's not as explicit
what's happening, but it should still be clear to most people reading
the resume what those sections are.