Add revision notes, test script for executable bit
Originally written via Gist in January 2015: I asked for a review of the
work I had done so far, and the notes that came back were hand-written
on a printed copy. In order to not lose them, I added them in as
comments to the file before I attempted to incorporate the changes.

Additionally, as I was having trouble with the script to generate the
PDF, I included a new command to test the ability to execute the details
script before trying to do so.