Revise details script tests, tweak page format
Originally written via Gist in January 2015: The details script tests
were rewritten to be more idiomatic for shell script, and mostly the
formatting of the format file itself was changed. However, I did attempt
to make it so that the details section of the script would collapse the
blank line, if additional details such as address and phone number were
not made available. This change was incomplete; I could never quite get
it to work like I hoped.