Vendor "xargs.sty"
Because I'm angling for using a Dockerfile to simplify the instructions
required to build the PDF, I've hit a few problems along the way. One in
particular that I can't solve any other (easy) way is the presence of
the "xargs" package. I'm going to be using a CentOS 7 base image, and
while most common CTAN packages are available through yum via
"texlive-$pkgname", "xargs" is not one of them. I could download it
through CTAN, but there's also a concern that it could change over time,
breaking the compilation unexpectedly. Plus, I prefer not to download
into the /resume directory during the build step for various reasons.

In any case, the most simple solution is to "vendor" this code, after a
fashion, by merely including it in the repository. It's kind of a hack,
and will result in an interesting licensing situation, but it should
work.