Add Dockerfile to build the resume
Now it's possible to build this resume by using Docker to create an
image capable of assembling the LaTeX files into a PDF. Using this
Dockerfile, it's possible to build an image, and then run it in a
one-off capacity, outputting the results into the `outputs/` directory,
of which the main file we care about (the resume itself), is symlinked
and availble at `resume.pdf`.

For example, these two commands should now be sufficient to build the
resume entirely:

    docker build -t resume .
    docker run --rm -v "$(pwd)/outputs:/resume/outputs" resume

Theoretically this only ever touches the "outputs/" subdirectory on disk
and so is impossible to modify the repository unexpectedly. However, if
desired (for development purposes), the entire repository can be
mounted so that any changes are immediately visible without having to
re-run the first `build` step. Instead, simply replace the second
command with the following:

    docker run --rm -v "$(pwd):/resume" resume

Eventually the above documentation should make its way into a `README`
file, but for now that should be enough to go on.