Remove some Vim-provided warnings
Technically, the "." after a series of capital letters is treated as an
end to an abbreviation, which winds up with slightly-different kerning
than the period indicating the end of a sentence. Using "\@." instead of
a bare "." fixes that. It reduces readability slightly, which makes it
work slightly against the desired behavior to be immediately-legible to
even non-LaTeX users, sometimes slight abberrations like this are
unavoidable when using a formatting language.

Vim also warned against having backslash commands such as \repo and
\LaTeX terminated by a space. The former is fixed by removing the extra
space, which was only there for aligning purposes (and this was
unnecessary as there was only one other element that would match it),
and the other is fixed by adding an additional backslash after \LaTeX,
before the space itself.