Add README and a license
The README file now contains instructions for how to build the project.
It is unfortunately not a particularly quick and lightweight process,
but it is a simple one to follow.

The license applies to most files except for the `resume.tex` file
(which is largely prose content and I do not want people to feel free to
alter it in minor ways), and the `xargs.sty` file (which I did not write
and do not have the authority to choose or alter licensing terms for).
The README file also describes these ideas in detail.